Conflicts:
	kalite/templates/base.html
	kalite/templates/video.html

+++ start.sh
@@ -1,9 +1,18 @@
+#!/bin/bash
+
 cd `dirname "${BASH_SOURCE[0]}"`/kalite
 
-# move any previously downloaded content from the old location to the new
-mv static/videos/* ../content > /dev/null 2> /dev/null
+if [ -f "database/data.sqlite" ] ; then
+	
+    # move any previously downloaded content from the old location to the new
+	mv static/videos/* ../content > /dev/null 2> /dev/null
 
-echo
-./cronstart.sh
-echo
-./serverstart.sh
+	echo
+	./cronstart.sh
+	echo
+	./serverstart.sh
+else
+	echo "Please run install.sh first!"
+fi
+
+cd ..
